+++ attributes/default.rb
@@ -21,4 +21,7 @@ default['owi_docker']['machine']['version'] = '0.13.0'
 # Specify the version of Docker Compose required. Otherwise, use the
 # `default['owi_docker']['compose'][binary_location]` to specify your own binary
 # location
-default['owi_docker']['compose']['version'] = '1.18.0'
+default['owi_docker']['compose']['version'] = '1.19.0'
+
+default['owi_docker']['service'] = {}
+default['owi_docker']['container'] = {}