+++ CHANGELOG.md
@@ -3,7 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.6] - 2017-00-02
+## [0.0.7] - 2017-01-02
+### Added
+- isuftin@usgs.gov - Docker Image wrapper recipe
+- isuftin@usgs.gov - Docker Container wrapper recipe
+### Updated
+- isuftin@usgs.gov - Kitchen configuration and styling
+
+## [0.0.6] - 2017-01-02
 ### Added
 - isuftin@usgs.gov - Added iptables recipe to allow implementors to update IPTables to allow Docker communication
 
@@ -14,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.0.4] - 2017-09-27
 ### Added
 - isuftin@usgs.gov - Added attributes into the Docker cookbook that match the upstream Docker cookbook
- 
+
 ## [0.0.3] - 2017-07-07
 ### Added
 - isuftin@usgs.gov - Fixed a very glaring bug in docker service creation